+++ src/block/block.integrational.spec.jsx
@@ -1,40 +1,52 @@
 import React from 'react';
 import ReactShallowRenderer from 'react-test-renderer/shallow';
+import isFunction from 'lodash/isFunction';
 import {Config} from '../config';
 import {block} from './block';
 import {modifier} from '../modifier';
 
 Config.ASSERTION_ENABLED = true;
 
-describe('BEM block decorator with separate modifier components', () => {
-    let renderer;
+describe('BEM block decorator', () => {
     let Foo;
-    let FooXyzzy;
-    let FooPlugh;
     beforeEach(() => {
-        renderer = new ReactShallowRenderer();
         Foo = () => <div />;
-        FooXyzzy = () => <div />;
-        FooPlugh = () => <div />;
     });
 
-    it('should choose component according to modifier', () => {
-        const WrappedFoo = block(
-            'foo',
-            ({xyzzy, plugh}) => [{xyzzy}, `plugh-${plugh}`]
-        )(
-            Foo,
-            modifier('xyzzy')(FooXyzzy),
-            modifier(/^plugh-\w\w$/)(FooPlugh)
-        );
-        renderer.render(<WrappedFoo />);
-        let wrappedFoo = renderer.getRenderOutput();
-        expect(wrappedFoo.type).toBe(Foo);
-        renderer.render(<WrappedFoo xyzzy />);
-        wrappedFoo = renderer.getRenderOutput();
-        expect(wrappedFoo.type).toBe(FooXyzzy);
-        renderer.render(<WrappedFoo plugh="xs" />);
-        wrappedFoo = renderer.getRenderOutput();
-        expect(wrappedFoo.type).toBe(FooPlugh);
+    it('should provide "namespaced" BEM element decorator', () => {
+        const WrappedFoo = block('foo')(Foo);
+        expect(isFunction(Foo.element)).toBeTruthy();
+        expect(isFunction(WrappedFoo.element)).toBeTruthy();
+    });
+
+    describe('with separate modifier components', () => {
+        let renderer;
+        let FooXyzzy;
+        let FooPlugh;
+        beforeEach(() => {
+            renderer = new ReactShallowRenderer();
+            FooXyzzy = () => <div />;
+            FooPlugh = () => <div />;
+        });
+
+        it('should choose component according to modifier', () => {
+            const WrappedFoo = block(
+                'foo',
+                ({xyzzy, plugh}) => [{xyzzy}, `plugh-${plugh}`]
+            )(
+                Foo,
+                modifier('xyzzy')(FooXyzzy),
+                modifier(/^plugh-\w\w$/)(FooPlugh)
+            );
+            renderer.render(<WrappedFoo />);
+            let wrappedFoo = renderer.getRenderOutput();
+            expect(wrappedFoo.type).toBe(Foo);
+            renderer.render(<WrappedFoo xyzzy />);
+            wrappedFoo = renderer.getRenderOutput();
+            expect(wrappedFoo.type).toBe(FooXyzzy);
+            renderer.render(<WrappedFoo plugh="xs" />);
+            wrappedFoo = renderer.getRenderOutput();
+            expect(wrappedFoo.type).toBe(FooPlugh);
+        });
     });
 });