+++ src/block/index.js
@@ -1,3 +1,2 @@
 export * from './block';
-export * from './block-context-types';
 export * from './plain-block';