+++ src/block/is-block-definition.js
@@ -1,0 +1,3 @@
+export function isBlockDefinition(Component) {
+    return Component && /^block\(.+\)$/.test(Component.displayName);
+}