+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.3] - 2017-05-26
+
+### Added
+
+- Namespaced *\@element* decorator provided by *\@block* and *\@plainBlock* decorators (see Issue \#9)
+
+### Fixed
+
+- Issue \#9 Element catches wrong block
+
 ## [1.0.2] - 2017-05-24
 
 ### Added
@@ -12,7 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-- Issue #8 Support mixins
+- Issue \#8 Support mixins
 
 ## [1.0.1] - 2017-05-19
 
@@ -41,3 +51,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [Unreleased]: https://github.com/redneckz/react-bem-helper/compare/v1.0.1...HEAD
 [1.0.1]: https://github.com/redneckz/react-bem-helper/compare/v1.0.0...v1.0.1
 [1.0.2]: https://github.com/redneckz/react-bem-helper/compare/v1.0.1...v1.0.2
+[1.0.3]: https://github.com/redneckz/react-bem-helper/compare/v1.0.2...v1.0.3